LayoutTests/http/tests/security/contentSecurityPolicy/video-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load media from 'http://localhost:8000/resources/test.mp4' because it violates the following Content Security Policy directive: "media-src http://127.0.0.1:8000/resources/redirect.php".

ALERT: PASS
